vehicle_track/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycrypto==2.6.1
requests==2.23.0
urllib3==1.25.9
Werkzeug==1.0.0
